requirements.txt
mysql-connector-python==9.2.0
pillow==11.1.0
prompt_toolkit==3.0.50
pyzmq==26.2.1
questionary==2.1.0
wcwidth==0.2.13
